Update match_fee type

+++ data/db_00119.sql
@@ -1,1 +1,1 @@
-ALTER TABLE `matches` CHANGE `prize_pool` `match_fee` INT;
+ALTER TABLE `matches` CHANGE `prize_pool` `match_fee` FLOAT;
